pyproject.toml
[project]
name = "sec-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.31",
    "google-cloud-storage>=3.8.0",
    "pandas>=2.3.3",
    "pyarrow>=23.0.0",
    "requests>=2.32.5",
]
